# shuf: improve performance