This PR improves the performance of `shuf` by replacing the current implementation of splitting input by a given separator with a version using the memchr crate. This crate was not used in `shuf` before, but it is already used in e.g. sort. Considering the benchmark results (see below), adding a new dependency to `shuf` is justified.

## Benchmarks

### Description

Generate input data (50,000,000 lines, 3.6 GiB):
```
cat /dev/random | base64 | fold | head -n 50000000 > input.txt
```
Run `shuf` with [hyperfine](https://github.com/sharkdp/hyperfine):
```
hyperfine -w 3 "target/release/coreutils shuf ../testdata/input.txt > /dev/null"
```
### Results

The input data was stored in tmpfs to avoid distortions from IO.

| Version  |  Time |
|---|---|
| Old version |  6.982 s ±  0.033 s |
| Improved version |  5.234 s ±  0.046 s |
| GNU shuf (v8.32) | 9.810 s ±  0.019 s |

The improved version is 25% faster than the previous version. Compared to the GNU implementation, `shuf` now is almost twice as fast.