@sylvestre 2022-06-16
Cool stuff. Could you please create a src/uu/shuf/BENCHMARKING.md  file ?
(see src/uu/*/BENCHMARKING.md  for example)