@sylvestre 2022-06-17
Some more benchmarks:
```
$ wget -O shakespeare.txt https://www.gutenberg.org/files/100/100-0.txt
$ hyperfine   -w 3 "./target/release/coreutils shuf shakespeare.txt > /dev/null" "shuf shakespeare.txt > /dev/null"              
Benchmark 1: ./target/release/coreutils shuf shakespeare.txt > /dev/null
  Time (mean ± σ):      22.1 ms ±   4.6 ms    [User: 15.3 ms, System: 6.8 ms]
  Range (min … max):    10.7 ms …  28.4 ms    114 runs
 
Benchmark 2: shuf shakespeare.txt > /dev/null
  Time (mean ± σ):      27.8 ms ±   7.4 ms    [User: 22.1 ms, System: 5.7 ms]
  Range (min … max):    14.5 ms …  42.1 ms    72 runs
 
Summary
  './target/release/coreutils shuf shakespeare.txt > /dev/null' ran
    1.26 ± 0.42 times faster than 'shuf shakespeare.txt > /dev/null'
```

=> faster here

But slower with:
```
$ wget https://www.gutenberg.org/files/30613/30613-0.txt -O odyssey.txt
$ hyperfine --shell=none  -w 3 "./target/release/coreutils shuf odyssey.txt " "shuf odyssey.txt "
Benchmark 1: ./target/release/coreutils shuf odyssey.txt 
  Time (mean ± σ):       3.4 ms ±   0.3 ms    [User: 1.6 ms, System: 1.6 ms]
  Range (min … max):     2.0 ms …   3.8 ms    866 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Benchmark 2: shuf odyssey.txt 
  Time (mean ± σ):       2.2 ms ±   0.3 ms    [User: 1.5 ms, System: 0.5 ms]
  Range (min … max):     1.1 ms …   3.4 ms    1209 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Summary
  'shuf odyssey.txt ' ran
    1.53 ± 0.24 times faster than './target/release/coreutils shuf odyssey.txt '
```