@anastygnome 2022-06-17
`bash

❯ hyperfine --shell=none  -w 3 "./target/release/coreutils shuf /tmp/shakespeare.txt " "shuf /tmp/shakespeare.txt "

Benchmark 1: ./target/release/coreutils shuf /tmp/shakespeare.txt 
  Time (mean ± σ):      15.9 ms ±   5.5 ms    [User: 11.3 ms, System: 4.4 ms]
  Range (min … max):    10.1 ms …  24.4 ms    281 runs
 
Benchmark 2: shuf /tmp/shakespeare.txt 
  Time (mean ± σ):      21.8 ms ±   6.7 ms    [User: 17.2 ms, System: 4.4 ms]
  Range (min … max):    14.8 ms …  36.1 ms    193 runs
 
Summary
  './target/release/coreutils shuf /tmp/shakespeare.txt ' ran
    1.37 ± 0.64 times faster than 'shuf /tmp/shakespeare.txt '

❯ hyperfine --shell=none  -w 3 "./target/release/coreutils shuf /tmp/odyssey.txt " "shuf /tmp/odyssey.txt "

Benchmark 1: ./target/release/coreutils shuf /tmp/odyssey.txt 
  Time (mean ± σ):       1.4 ms ±   0.5 ms    [User: 0.9 ms, System: 0.4 ms]
  Range (min … max):     0.8 ms …   2.5 ms    1527 runs
 
Benchmark 2: shuf /tmp/odyssey.txt 
  Time (mean ± σ):       1.4 ms ±   0.5 ms    [User: 0.9 ms, System: 0.4 ms]
  Range (min … max):     0.8 ms …   2.4 ms    1729 runs
 
Summary
  'shuf /tmp/odyssey.txt ' ran
    1.02 ± 0.53 times faster than './target/release/coreutils shuf /tmp/odyssey.txt '

`

@sylvestre nice I/O benchmark of yours there ;)

You should put the file in tmpfs to negate this issue (in /tmp/ in my case)