@sylvestre 2022-06-17
Cool :)
Just create the benchmark doc and it can land!