@anastygnome 2022-06-17
> oups, sorry :) Please let @Garfield96 do it :)

On a side note, please look at the following markdown syntax for CLI output 
```markdown

```console
$ whoami 
anastygnome 
<3x ` to END>
```
yields 

```console
$ whoami 
anastygnome 
```