@Garfield96 2022-06-17
Thanks, @sylvestre, and @anastygnome for doing additional benchmarks. I also tested `https://www.gutenberg.org/files/30613/30613-0.txt` and I can confirm that the GNU version in some runs showed a slightly better performance. Since this input is very small, applying the shuffling does not dominate the execution time. Instead, loading the binary becomes the bottleneck. While the combined build of coreutils always produced worse results on my system (1.2ms compared to 0.9ms for GNU shuf), building `shuf` as a standalone binary solved the problem (runtimes were equal).

I extended the description in `BENCHMARKING.md`, hence this PR should be complete.